Fix failing to find libc on musl Linux if gcc is not installed.
ctypes.util.find_library() internally uses ldconfig, then if that does not work,
it switches to using objdump and then gcc. The musl variant of ldconfig does not
work for find_library() causing find_library('c') to default to using gcc if
available or not finding it at all if it isn't.

This broke the requirement that gcc should be a build time dependency only.