Expose all stdlib functions available on the current operating system.
The original plan of making cslug.stdlib expose only symbols that are
universally available on all platforms meant that, as cslug gained new platforms
or newer versions of existing platforms dropped functions (in particularly the
dropping of the insecure family of functions that took NULL terminated string
inputs), that the pool of functions cslug could expose was constantly shrinking
and cslug was essentially having to make breaking changes every time.

The new approach is simply to expose anything that's available on the current
platform then document which functions are available on each platform.