stdlib: Use new universal C runtime on Windows if available.
Windows 11 (and optionally newer versions of Windows 10 a certain update is
installed) introduces yet another libc implementation - the Universal C Runtime
(UCRT). Unlike msvcrt.dll which cslug currently uses, it receives updates and is
officially documented and supported API so it shouldn't disappear at some
unknown point in the future.

If this library is unavailable (i.e. on an older Windows), msvcrt.dll is used as
before.