Enforce macOS deployment target.
The -mmacosx-version-min flag doesn't do anything to make binaries compatible
with the version declared. Nor does it do anything to detect if the binaries the
compiler produces are compatible with the deployment target unless some extra
magic flags are added to opt into ensuring an executable is functional 🙄.

Add those extra flags.